% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamCheck.R
\name{bamCheck}
\alias{bamCheck}
\title{bamCheck}
\usage{
bamCheck(x = NULL, filetype = NULL, outname = NULL)
}
\arguments{
\item{x}{file path to BAM/CRAM file}

\item{filetype}{character string matching either "BAM" or "CRAM"}

\item{outname}{file path for output}
}
\value{
none
}
\description{
bamCheck
}
