% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridStats.R
\name{depthtocn}
\alias{depthtocn}
\title{depthtocn}
\usage{
depthtocn(x, purity, seqdepth)
}
\arguments{
\item{x}{numeric vector}

\item{purity}{float}

\item{seqdepth}{integer}
}
\value{
numeric vector
}
\description{
converts read depth to copy number given purity and single copy depth
}
