% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsFittingColumnNames}
\alias{dsFittingColumnNames}
\title{dsFittingColumnNames}
\format{
character vector
}
\usage{
dsFittingColumnNames
}
\description{
dsFittingColumnNames
}
\keyword{datasets}
