% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFitTable.R
\name{filterFitTable}
\alias{filterFitTable}
\title{filterFitTable}
\usage{
filterFitTable(
  table = NULL,
  metadata = NULL,
  filter_underpowered = NULL,
  filter_homozygous = NULL,
  af_cutoff = NULL,
  ranks = 10
)
}
\arguments{
\item{table}{data.frame of absolute cn fits}

\item{metadata}{data.frame of workflow metadata}

\item{filter_underpowered}{Boolean to filter underpowered fits}

\item{filter_homozygous}{Boolean to filter based on homozygous loss}

\item{af_cutoff}{Float for TP53 anchoring difference threshold}

\item{ranks}{Integer for number of top fits to return}
}
\value{
data.frame
}
\description{
filterFitTable
}
