% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDownsampleDepth.R
\name{getDownsampleDepth}
\alias{getDownsampleDepth}
\title{getDownsampleDepth}
\usage{
getDownsampleDepth(
  ploidy = NULL,
  purity = NULL,
  nbins_ref_genome = NULL,
  rpc = 15,
  ratio = 1.098901
)
}
\arguments{
\item{ploidy}{numeric ploidy value}

\item{purity}{numeric purity value}

\item{nbins_ref_genome}{number of bins in reference genome}

\item{rpc}{minimum read count per bin per copy}

\item{ratio}{availble/used read ratio}
}
\value{
numeric
}
\description{
getDownsampleDepth
}
