% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneSeg.R
\name{getGeneSeg}
\alias{getGeneSeg}
\title{getGeneSeg}
\usage{
getGeneSeg(target = NULL, abs_data = NULL, genome = NULL)
}
\arguments{
\item{target}{character string of genomic region in format "N:start-end"}

\item{abs_data}{QDNASeq object of class "QDNAseqCopyNumbers"}

\item{genome}{reference genome}
}
\value{
numeric vector
}
\description{
getGeneSeg
}
