% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridStats.R
\name{gridStats}
\alias{gridStats}
\title{gridStats}
\usage{
gridStats(obj = NULL, ploidy = NULL, purity = NULL)
}
\arguments{
\item{obj}{QDNASeq object of class "QDNAseqCopyNumbers"}

\item{ploidy}{numeric ploidy value}

\item{purity}{numeric purity value}
}
\value{
list
}
\description{
gridStats
}
