% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{plotProfile}
\alias{plotProfile}
\title{plotProfile}
\usage{
plotProfile(
  relcn = NULL,
  ploidy = NA,
  purity = NA,
  clonality = NA,
  rmse = NA,
  yrange = NULL
)
}
\arguments{
\item{relcn}{QDNAseq object}

\item{ploidy}{numeric ploidy}

\item{purity}{numeric purity}

\item{clonality}{MAE value}

\item{rmse}{RMSE value}

\item{yrange}{max y-axis value}
}
\value{
plot
}
\description{
plotProfile
}
