% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictProfile.R
\name{predictProfile}
\alias{predictProfile}
\title{predictProfile}
\usage{
predictProfile(
  qctable = NULL,
  model = NULL,
  method = "randforest",
  flagThreshold = 0.74,
  errorMetric = "clonality"
)
}
\arguments{
\item{qctable}{data.frame of absolute fits}

\item{model}{ML model object}

\item{method}{Method for autofitting}

\item{flagThreshold}{Threshold for flagging fits}

\item{errorMetric}{Error metric for fitting}
}
\value{
Data.frame of fitted absolute fits
}
\description{
predictProfile
}
