% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswgsabsolutecn-package.R
\docType{package}
\name{rswgsabsolutecn-package}
\alias{rswgsabsolutecn}
\alias{rswgsabsolutecn-package}
\title{rswgsabsolutecn: R package for functions invcluded in the swgs-absolutecn workflow}
\description{
R functions used by swgs-absolutecn workflow
}
\author{
\strong{Maintainer}: Philip Smith \email{philip.smith@cruk.cam.ac.uk}

}
\keyword{internal}
