% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothSample.R
\name{smoothSample}
\alias{smoothSample}
\title{smoothSample}
\usage{
smoothSample(relcn = NULL, smooth = FALSE, maxSegs = 300, seed = NULL)
}
\arguments{
\item{relcn}{QDNASeq object}

\item{smooth}{Boolean to apply smoothing}

\item{maxSegs}{Maximum segments}

\item{seed}{Value for seed}
}
\value{
QDNASeq object
}
\description{
Adjusted recursively to drop max segments below threshold using maximum StdDev
difference in means during segmentation splits in CBS
}
