% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swgs_abs_rfranger_model}
\alias{swgs_abs_rfranger_model}
\title{swgs_abs_rfranger_model
Ranger randomforest model used in autofitting trained
using 900 fitted profiles}
\format{
character vector
}
\usage{
swgs_abs_rfranger_model
}
\description{
swgs_abs_rfranger_model
Ranger randomforest model used in autofitting trained
using 900 fitted profiles
}
\keyword{datasets}
